lib: Create the `simple_workspace_store` module
This is in preparation of just moving the `WorkspaceStore` trait to `jj-core`.
