+++ cli/src/commands/workspace/forget.rs
@@ -19,7 +19,7 @@ use jj_lib::git::GitSubprocessOptions;
 use jj_lib::ref_name::WorkspaceNameBuf;
 #[cfg(feature = "git")]
 use jj_lib::repo::Repo as _;
-use jj_lib::workspace_store::SimpleWorkspaceStore;
+use jj_lib::simple_workspace_store::SimpleWorkspaceStore;
 use jj_lib::workspace_store::WorkspaceStore as _;
 use tracing::instrument;
 
@@ -78,7 +78,7 @@ pub async fn cmd_workspace_forget(
         return Ok(());
     }
 
-    let workspace_store = SimpleWorkspaceStore::load(workspace_command.repo_path())?;
+    let simple_workspace_store = SimpleWorkspaceStore::load(workspace_command.repo_path())?;
 
     #[cfg(feature = "git")]
     let workspace_paths = {
@@ -86,7 +86,10 @@ pub async fn cmd_workspace_forget(
         forget_ws
             .iter()
             .filter_map(|ws| {
-                let rel_path = workspace_store.get_workspace_path(ws).ok().flatten()?;
+                let rel_path = simple_workspace_store
+                    .get_workspace_path(ws)
+                    .ok()
+                    .flatten()?;
                 dunce::canonicalize(repo_path.join(rel_path)).ok()
             })
             .collect_vec()
@@ -100,7 +103,7 @@ pub async fn cmd_workspace_forget(
         tx.repo_mut().remove_workspace(ws).await?;
     }
 
-    workspace_store.forget(&forget_ws.iter().map(|x| x.as_ref()).collect::<Vec<_>>())?;
+    simple_workspace_store.forget(&forget_ws.iter().map(|x| x.as_ref()).collect::<Vec<_>>())?;
 
     let description = if let [ws] = forget_ws.as_slice() {
         format!("forget workspace {}", ws.as_symbol())
