+++ cli/src/commands/workspace/rename.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use jj_lib::ref_name::WorkspaceNameBuf;
-use jj_lib::workspace_store::SimpleWorkspaceStore;
+use jj_lib::simple_workspace_store::SimpleWorkspaceStore;
 use jj_lib::workspace_store::WorkspaceStore as _;
 use tracing::instrument;
 
@@ -60,7 +60,7 @@ pub async fn cmd_workspace_rename(
         )));
     }
 
-    let workspace_store = SimpleWorkspaceStore::load(workspace_command.repo_path())?;
+    let simple_workspace_store = SimpleWorkspaceStore::load(workspace_command.repo_path())?;
 
     let mut tx = workspace_command.start_transaction().into_inner();
     let (mut locked_ws, _wc_commit) = workspace_command.start_working_copy_mutation().await?;
@@ -70,7 +70,7 @@ pub async fn cmd_workspace_rename(
     tx.repo_mut()
         .rename_workspace(&old_name, new_name.to_owned())?;
 
-    workspace_store.rename(&old_name, new_name)?;
+    simple_workspace_store.rename(&old_name, new_name)?;
 
     let repo = tx
         .commit(format!(
